requirements.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dronekit==2.9.2
fonttools==4.53.1
future==1.0.0
gcloud==0.18.3
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
importlib_resources==6.4.5
jwcrypto==1.5.6
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
monotonic==1.6
numpy==2.0.2
oauth2client==4.1.3
packaging==24.1
pillow==10.4.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pymavlink==2.4.41
pyparsing==3.1.4
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-jwt==4.1.0
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
typing_extensions==4.12.2
urllib3==1.26.20
zipp==3.20.2
